pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codeevolver-agents"
version = "0.1.0"
description = "Remote service for executing evolutionary code changes"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "motor>=3.6.0",
    "pydantic-settings>=2.6.0",
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/codeevolver_agents"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
